Organize tests a bit better
Split up the gigantic SerializationTest into easier-to-maintain test cases that are more behavior focused.
