+++ src/leapserial/test/SerializationTest.cpp
@@ -14,18 +14,13 @@ class SerializationTest:
 static_assert(leap::has_serializer<std::string>::value, "std::string");
 static_assert(leap::has_serializer<std::vector<int>>::value, "Vector of ints");
 static_assert(leap::has_serializer<std::vector<std::string>>::value, "Vector of strings");
-static_assert(leap::has_serializer<std::map<int, int>>::value, "Map of primitives");
-static_assert(leap::has_serializer<std::map<std::string, std::string>>::value, "Map of nonprimitives");
-static_assert(leap::has_serializer<std::unordered_map<int, int>>::value, "Unordered map of primitives");
-static_assert(leap::has_serializer<std::unordered_map<std::string, std::string>>::value, "Unordered map of nonprimitives");
+static_assert(leap::serializer_is_irresponsible<std::vector<int*>>::value, "Irresponsible allocator was not correctly inferred");
+static_assert(!leap::serializer_is_irresponsible<std::vector<int>>::value, "Irresponsibility incorrectly inferred in a responsible allocator");
 
 struct DummyStruct {};
 static_assert(!leap::has_serializer<DummyStruct>::value, "A dummy structure was incorrectly classified as having a valid serializer");
 static_assert(!leap::has_serializer<DummyStruct[2]>::value, "An arry of dummy structures was incorrectly classified as having a valid serializer");
 static_assert(!leap::has_serializer<std::vector<DummyStruct>>::value, "A vector of dummy structures was incorrectly classified as having a valid serializer");
-static_assert(!leap::has_serializer<std::unique_ptr<DummyStruct>>::value, "A unique pointer to a dummy structure was incorrectly classified as having a valid serializer");
-static_assert(!leap::has_serializer<std::map<int, DummyStruct>>::value, "A map of DummyStruct was incorrectly classified as having a valid serializer");
-static_assert(!leap::has_serializer<std::unordered_map<int, DummyStruct>>::value, "An unordered_map of DummyStruct was incorrectly classified as having a valid serializer");
 static_assert(!leap::has_serializer<std::vector<DummyStruct>>::value, "A vector of DummyStruct was incorrectly classified as having a valid serializer");
 
 struct MyRepeatedStructure {
@@ -144,178 +139,6 @@ TEST_F(SerializationTest, VerifyRoundTrip) {
 
   ASSERT_STREQ("Hello world!", allocation->myString.c_str()) << "Standard string round trip serialization did not work";
   ASSERT_STREQ(L"World hello!", allocation->myWString.c_str()) << "Standard string round trip serialization did not work";
-}
-
-struct SimpleStruct {
-  std::map<int, int> myMap;
-  std::unordered_map<int, int> myUnorderedMap;
-
-  static leap::descriptor GetDescriptor(void) {
-    return {
-      &SimpleStruct::myMap,
-      &SimpleStruct::myUnorderedMap
-    };
-  }
-};
-
-TEST_F(SerializationTest, CanSerializeMaps) {
-  std::stringstream ss;
-
-  {
-    SimpleStruct x;
-    x.myMap[0] = 555;
-    x.myMap[232] = 12;
-    x.myUnorderedMap[29] = 44;
-    x.myUnorderedMap[30] = 4014;
-    leap::Serialize(ss, x);
-  }
-
-  // Deserialize, make sure map entries are in the right place
-  auto x = leap::Deserialize<SimpleStruct>(ss);
-  ASSERT_EQ(2UL, x->myMap.size()) << "Map has an incorrect number of elements";
-  ASSERT_EQ(555, x->myMap[0]);
-  ASSERT_EQ(12, x->myMap[232]);
-
-  // Same for unordered map
-  ASSERT_EQ(2UL, x->myUnorderedMap.size()) << "Unordered map has an incorrect number of elements";
-  ASSERT_EQ(44, x->myUnorderedMap[29]);
-  ASSERT_EQ(4014, x->myUnorderedMap[30]);
-}
-
-template<typename T>
-struct MyLinkedList {
-  MyLinkedList(const T& v = T()) : value(v), prev(nullptr), next(nullptr) {}
-  void InsertAfter(MyLinkedList* last) {
-    if (!last)
-      return;
-    prev = last;
-    next = prev->next;
-    if (next) {
-      next->prev = this;
-    }
-    prev->next = this;
-  }
-
-  T value;
-  MyLinkedList* prev;
-  MyLinkedList* next;
-
-  static leap::descriptor GetDescriptor(void) {
-    return {
-      &MyLinkedList::value,
-      &MyLinkedList::prev,
-      &MyLinkedList::next
-    };
-  }
-};
-
-TEST_F(SerializationTest, VerifyDoublyLinkedList) {
-  std::stringstream ss;
-
-  {
-    MyLinkedList<int> head{1};
-    MyLinkedList<int> middle{2};
-    MyLinkedList<int> tail{3};
-    middle.InsertAfter(&head);
-    tail.InsertAfter(&middle);
-    // Doesn't support null right now, so add loops
-    head.prev = &tail;
-    tail.next = &head;
-
-    // Verfiy that we are creating what we think we are creating:
-    ASSERT_EQ(1, head.value);
-    ASSERT_TRUE(head.prev != nullptr);
-    ASSERT_TRUE(head.next != nullptr);
-    ASSERT_EQ(3, head.prev->value);
-    ASSERT_EQ(2, head.next->value);
-    ASSERT_EQ(2, middle.value);
-    ASSERT_TRUE(middle.prev != nullptr);
-    ASSERT_TRUE(middle.next != nullptr);
-    ASSERT_EQ(3, middle.next->value);
-    ASSERT_EQ(1, middle.prev->value);
-    ASSERT_TRUE(tail.prev != nullptr);
-    ASSERT_TRUE(tail.next != nullptr);
-    ASSERT_EQ(3, tail.value);
-    ASSERT_EQ(1, tail.next->value);
-    ASSERT_EQ(2, tail.prev->value);
-
-    leap::Serialize(ss, head);
-  }
-
-  auto ll = leap::Deserialize<MyLinkedList<int>>(ss);
-  MyLinkedList<int>& head = *ll;
-  ASSERT_TRUE(head.next != nullptr);
-  MyLinkedList<int>& middle = *head.next;
-  ASSERT_TRUE(middle.next != nullptr);
-  MyLinkedList<int>& tail = *middle.next;
-
-  ASSERT_EQ(1, head.value);
-  ASSERT_TRUE(head.next != nullptr);
-  ASSERT_TRUE(head.prev != nullptr);
-  ASSERT_EQ(3, head.prev->value);
-  ASSERT_EQ(2, head.next->value);
-  ASSERT_EQ(2, middle.value);
-  ASSERT_TRUE(middle.prev != nullptr);
-  ASSERT_TRUE(middle.next != nullptr);
-  ASSERT_EQ(3, middle.next->value);
-  ASSERT_EQ(1, middle.prev->value);
-  ASSERT_EQ(3, tail.value);
-  ASSERT_TRUE(tail.prev != nullptr);
-  ASSERT_TRUE(tail.next != nullptr);
-  ASSERT_EQ(1, tail.next->value);
-  ASSERT_EQ(2, tail.prev->value);
-}
-
-TEST_F(SerializationTest, NullPointerDereferenceCheck) {
-  std::stringstream ss;
-  {
-    MyLinkedList<int> ll;
-    ll.next = &ll;
-    ll.prev = nullptr;
-    ll.value = 192;
-
-    leap::Serialize(ss, ll);
-  }
-
-  // Deserialize, ensure we get null back
-  {
-    auto ll = leap::Deserialize<MyLinkedList<int>>(ss);
-    ASSERT_EQ(192, ll->value);
-    ASSERT_EQ(ll.get(), ll->next);
-    ASSERT_EQ(nullptr, ll->prev);
-  }
-}
-
-struct StructureWithEmbeddedField {
-  MyLinkedList<int> embedded;
-
-  static leap::descriptor GetDescriptor(void) {
-    return {
-      &StructureWithEmbeddedField::embedded
-    };
-  }
-};
-
-TEST_F(SerializationTest, EmbeddedFieldCheck) {
-  StructureWithEmbeddedField root;
-  MyLinkedList<int> top;
-
-  root.embedded.value = 1029;
-  root.embedded.prev = &root.embedded;
-  root.embedded.next = &top;
-
-  top.value = 292;
-  top.prev = &root.embedded;
-  top.next = nullptr;
-
-  // Round-trip serialization:
-  std::stringstream ss;
-  leap::Serialize(ss, root);
-
-  auto swef = leap::Deserialize<StructureWithEmbeddedField>(ss);
-
-  ASSERT_EQ(1029, swef->embedded.value);
-  ASSERT_EQ(292, swef->embedded.next->value);
 }
 
 struct HasOneMember {
@@ -678,128 +501,6 @@ TEST_F(SerializationTest, CanSerializeBoolsTest) {
   }
 }
 
-static_assert(leap::serializer_is_irresponsible<int*>::value, "Irresponsible allocator was not correctly inferred");
-static_assert(leap::serializer_is_irresponsible<std::vector<int*>>::value, "Irresponsible allocator was not correctly inferred");
-static_assert(leap::serializer_is_irresponsible<std::map<std::string, std::vector<int*>>>::value, "Irresponsible allocator was not correctly inferred");
-static_assert(!leap::serializer_is_irresponsible<std::vector<int>>::value, "Irresponsibility incorrectly inferred in a responsible allocator");
-static_assert(!leap::serializer_is_irresponsible<std::map<std::string, std::vector<int>>>::value, "Irresponsibility incorrectly inferred in a responsible allocator");
-
-TEST_F(SerializationTest, ComplexMapOfStringToVectorTest) {
-  std::stringstream ss;
-
-  // Serialize the complex map type first
-  {
-    std::map<std::string, std::vector<int>> mm;
-    mm["a"] = std::vector<int> {1, 2, 3};
-    mm["b"] = std::vector<int> {4, 5, 6};
-    mm["c"] = std::vector<int> {7, 8, 9};
-
-    leap::Serialize(ss, mm);
-  }
-
-  // Try to get it back:
-  std::map<std::string, std::vector<int>> mm;
-  leap::Deserialize(ss, mm);
-
-  ASSERT_EQ(1UL, mm.count("a")) << "Deserialized map was missing an essential member";
-  ASSERT_EQ(1UL, mm.count("b")) << "Deserialized map was missing an essential member";
-  ASSERT_EQ(1UL, mm.count("c")) << "Deserialized map was missing an essential member";
-
-  const auto& a = mm["a"];
-  ASSERT_EQ(3UL, a.size()) << "Vector was sized incorrectly";
-
-  const auto& c = mm["c"];
-  ASSERT_EQ(3UL, c.size()) << "Vector was sized incorrectly";
-  ASSERT_EQ(7, c[0]);
-  ASSERT_EQ(8, c[1]);
-  ASSERT_EQ(9, c[2]);
-}
-
-struct HasThreeSharedPointers {
-  std::shared_ptr<int> a;
-  std::shared_ptr<int> b;
-  std::shared_ptr<int> c;
-
-  static leap::descriptor GetDescriptor(void) {
-    return {
-      &HasThreeSharedPointers::a,
-      &HasThreeSharedPointers::b,
-      &HasThreeSharedPointers::c
-    };
-  }
-};
-
-TEST_F(SerializationTest, CanSerializeSharedPtrs) {
-  HasThreeSharedPointers htsp;
-  htsp.a = std::make_shared<int>(101);
-  htsp.b = std::make_shared<int>(199);
-  htsp.c = htsp.a;
-
-  // Round trip
-  std::stringstream ss;
-  leap::Serialize(ss, htsp);
-
-  HasThreeSharedPointers res;
-  leap::Deserialize(ss, res);
-  ASSERT_EQ(res.a, res.c) << "Aliased shared pointer was not correctly detected in a second pass";
-  ASSERT_EQ(101, *res.a) << "Aliased shared pointer was not correctly stored";
-  ASSERT_EQ(199, *res.b) << "Ordinary shared pointer was not correctly stored";
-
-  ASSERT_EQ(2UL, res.a.use_count()) << "Incorrect use count detected";
-}
-
-TEST_F(SerializationTest, SharedPtrOfSharedPtrs) {
-  std::shared_ptr<HasThreeSharedPointers> sphtsp = std::make_shared<HasThreeSharedPointers>();
-  sphtsp->a = std::make_shared<int>(101);
-  sphtsp->b = std::make_shared<int>(199);
-  sphtsp->c = sphtsp->a;
-
-  // Round trip
-  std::stringstream ss;
-  leap::Serialize(ss, sphtsp);
-
-  std::shared_ptr<HasThreeSharedPointers> res;
-  leap::Deserialize(ss, res);
-  ASSERT_EQ(res->a, res->c) << "Aliased shared pointer was not correctly detected in a second pass";
-  ASSERT_EQ(101, *res->a) << "Aliased shared pointer was not correctly stored";
-  ASSERT_EQ(199, *res->b) << "Ordinary shared pointer was not correctly stored";
-
-  ASSERT_EQ(2UL, res->a.use_count()) << "Incorrect use count detected";
-}
-
-TEST_F(SerializationTest, VectorOfSharedPointers) {
-  std::vector<std::shared_ptr<int>> vosp;
-
-  // Fill the vector with originals:
-  for (int i = 10; i--;)
-    vosp.push_back(std::make_shared<int>(i));
-
-  // Create some duplicates:
-  vosp.resize(14);
-  vosp[10] = vosp[4];
-  vosp[11] = vosp[4];
-  vosp[12] = vosp[5];
-  vosp[13] = vosp[9];
-
-  // Round trip, as per usual
-  std::stringstream ss;
-  leap::Serialize(ss, vosp);
-
-  std::vector<std::shared_ptr<int>> ret;
-  leap::Deserialize(ss, ret);
-  ASSERT_EQ(14, ret.size()) << "Incorrect number of deserialized entries";
-
-  const char* msg = "Aliased entries did not correctly match when deserialized";
-  ASSERT_EQ(ret[4], ret[10]) << msg;
-  ASSERT_EQ(ret[4], ret[11]) << msg;
-  ASSERT_EQ(ret[5], ret[12]) << msg;
-  ASSERT_EQ(ret[9], ret[13]) << msg;
-
-  ASSERT_EQ(3UL, ret[4].use_count()) << "Expected three aliases when deserializing";
-  ASSERT_EQ(2UL, ret[5].use_count()) << "Expected two aliases when deserializing";
-  ASSERT_EQ(2UL, ret[9].use_count()) << "Expected two aliases when deserializing";
-}
-
 struct SimpleObject {
   SimpleObject(int val):
     i(val)
@@ -1129,33 +830,4 @@ TEST_F(SerializationTest, EnumClassTest) {
 
   auto x = leap::Deserialize<SimpleEnum>(ss);
   ASSERT_EQ(SimpleEnum::Four, *x) << "Round-trip serialization of an enum class yielded an incorrect result";
-}
-
-struct HasChronoMembers {
-  std::chrono::seconds one{ 4929 };
-  std::chrono::nanoseconds two{ 4929 };
-  std::chrono::duration<short, std::ratio<24 * 60 * 60, 1>> three{ 2 };
-  std::chrono::duration<float> four{ 1.2f };
-
-  static leap::descriptor GetDescriptor(void) {
-    return{
-      &HasChronoMembers::one,
-      &HasChronoMembers::two,
-      &HasChronoMembers::three,
-      &HasChronoMembers::four
-    };
-  }
-};
-
-TEST_F(SerializationTest, DurationMembers) {
-  const HasChronoMembers hcm{};
-
-  std::stringstream ss;
-  leap::Serialize(ss, hcm);
-  auto re = leap::Deserialize<HasChronoMembers>(ss);
-
-  ASSERT_EQ(hcm.one, re->one) << "Seconds did not deserialize properly";
-  ASSERT_EQ(hcm.two, re->two) << "Nanoseconds did not deserialize properly";
-  ASSERT_EQ(hcm.three, re->three) << "Custom days duration did not deserialize properly";
-  ASSERT_EQ(hcm.four, re->four) << "Floating point duration did not deserialize properly";
 }
